upload master layout

+++ src/app/Modules/facilities/Services/facilites.service.ts
@@ -10,20 +10,20 @@ export class FacilitesService {
 
   constructor(private _HttpClient:HttpClient) { }
     CreateFacility(data: any): Observable<IFaciltyModal> {
-    return this._HttpClient.post<IFaciltyModal>("room-facilities", data);
+    return this._HttpClient.post<IFaciltyModal>("admin/room-facilities", data);
   }
 
   EditFacility(id: string, data: any): Observable<IFaciltyModal> {
-    return this._HttpClient.put<IFaciltyModal>(`room-facilities/${id}`, data);
+    return this._HttpClient.put<IFaciltyModal>(`admin/room-facilities/${id}`, data);
   }
 
     deletefacilities(id: string): Observable<any> {
-    return this._HttpClient.delete(`room-facilities/${id}`);
+    return this._HttpClient.delete(`admin/room-facilities/${id}`);
   }
 
   getAllFacilites(page: number,size: number): Observable<IApiResponse<IFacilitiesData>> {
     return this._HttpClient.get<IApiResponse<IFacilitiesData>>(
-      `room-facilities?page=${page}&size=${size}`
+      `admin/room-facilities?page=${page}&size=${size}`
     );
   }
 }
